MODFQMMGR-57 Fix f_unaccent() call to use index
This commit updates the drv_item_callnumber_location view to use the
mod_inventory_storage version of f_unaccent(), so that it can make
better use of indexes. In order to make this change, the commit also
moves that view definition from a SQL file to XML, so that we can easily
access the tenant name. The view was reformatting as part of that, but
the only substantial change is the different schema for f_unaccent()