MODFQMMGR-73 Purge old queries using the start date
This commit modifies the query purge process, to make it use the query
start date instead of the end date. If the server shuts down while a
query is still running, the end date doesn't get set, so those queries
weren't getting purged. Using the start date avoids that issue. This
means that queries may get purged a little sooner (e.g., if a query
takes 20 minutes to run, we'd potentially purge it 20 minutes sooner),
but the threshold is long enough (3 hours and can be overridden at
runtime) that this should be okay