MODFQMMGR-81 Refactor case sensitivity in queries
This commit refactors how we handle case-sensitivity when processing FQL
queries. With this change, when we do string comparisons, we always
convert the value being compared against to lower-case. The field is
only converted to lower-case if it doesn't have a filterValueGetter,
since the assumption is that the filterValueGetter will handle
converting the field to lower-case (if this assumotion changes, we will
need to update this behavior)

This commit also refactors FqlToSqlConverterService a bit, to cut down
on repeated work (before, nearly every method was recreating the Field
object for every FQL node/comparison in a query, which is a pretty
expensive operation. Now, this only happens once per node)