MODFQMMGR-110 Update a JOIN condition to use indexes
This commit updates the JOIN condition in the Loans entity type for the
user patron group table (src_users_groups), to use the
src_user_details.patrongroup instead of the value in
src_user_details.jsonb.

There are indexes available to use either the jsonb or the plain column,
but prior to this, the JOIN condition didn't match any of the indexes. I
opted to use the plain column index, since the index on the jsonb field
would require calling a function on a potentially large number of rows
in src_user_details, while the plain column requires no additional
processing