MODFQMMGR-178 Fix bug with regex+valueFunctions
This fixes a bug where valueFunctions were not getting used when
converting regex conditions to SQL, which could lead to unexpected
results.

Worth noting: this changes the behavior of regex conditions in a subtle
way that could lead to other expected results in the rare cases where
a field's valueFunction changes/breaks the semantics of a given regex.
This should be extremely rare, though, since the regex produced by the
query builder UI plugin only produces regular expressions that shouldn't
break in unexpected ways (and when the semantics does change, the
changes will generally be consistent with other operators)