MODFQMMGR-201 Add limit param to value source API calls
This commit adds a "limit" query parameter with a value of 1000 to every
value source API call, since some API endpoints have a fairly low
default limit. This brings the API approach to source values in line
with the DB approach (which also limits to 1000). In the future, we may
need to make this more flexible, in case we run into API endpoints in
FOLIO that reject this query parameter