MODFQMMGR-267 Only replace defined entity types when initializing
This commit changes the entity type initialization process, so that only
the entity types that are defined in JSON files are replaced. Any that
are in the DB, but aren't defined in JSON files, will be left untouched.
This prevents a non-deterministic test failure (in
FqlToSqlConverterServiceIT) and also allows for custom entity types