MODFQMMGR-408 Clean up EntityTypeFlatteningService
This commit updates the EntityTypeFlatteningService, so that it's not
allocating and destroying quite so many lists of columns. Instead, it
keeps everything in a Stream<Stream<EntityTypeColumn>>, which it then
flattens and reduces to a list when it's actually needed.

It also refactors getFlattenedEntityType() and splits it up a little
more, in order to reduce the congnitive complexity of the method

Finally, it fixes a broken test