MODFQMMGR-421 Make instance tenant_id not be ECS-only
This is a temporary change, to ensure the tenant ID is available as
part of the ID, even in non-ECS environments. After consumers have
switched away from the .../sortedIds endpoint, so that they can retrieve
the tenant ID more directly via query results, we can switch this back
to ECS-only.