MODFQMMGR-415 Add an entity type defintion cache
This commit adds a new cache for entity types to EntityTypeRepository,
to cut down on repeated loads of entity types. The default cache
duration is 1 hour, but this can be overridden via system property.
Additionally, the value in the module descriptor is set to 5 minutes, in
order to strike more of a balance between performance and practicality
in dev environments (although devs should consider reducing this further
when actively creating/updating entity type definitions)