MODFQMMGR-462 MODFQMMGR-477 Fix cross-tenant queries
This commit changes how FQM retrieves the tenant list to use for
cross-tenant queries. Now, it uses the mod-users' /user-tenants API to
determine if it's running in an ECS environment and retrieve the
consortia and central tenant IDs. Then, if it's an ECS environment, it
retrieves the actual list of tenant IDs for the user from
mod-consortia's user-tenants API.

Further, this changes how FQM determines the central tenant ID when
installing. First, it looks at the parameters provided when enabling the
module for the central tenant ID. If that isn't found, it falls back to
the mod-users /user-tenants API. The parameter should be used when
initially setting up an ECS environment, as the central tenant ID
generally isn't available from mod-users until after the modules have
been enabled. After the initial set up (e.g., when updating to a newer
module version), the central tenant ID is available from mod-users, so
the parameter is no longer needed.