MODFQMMGR-644 Add comment to composite_voucher_line_totals_per_account
composite_voucher_line_totals_per_account is an unusual "composite"
entity type, in that it is made up entirely of DB sources. Currently, we
are unable to update this to inherit its fields from other entity types
due to FQM's limited support for data aggregation. This commit adds a
comment to composite_voucher_line_totals_per_account.json5 to explain
this.