MODFQMMGR-543 Add support for ordering ET sources in composites
This commit makes it so columns within a composite are grouped together
in an order defined by the new "order" property on ET sources. Within
each group, the columns are sorted by labelAlias.

One major result of this is that in all cases, the columns in a
flattened entity type are always sorted list this; before they were only
sorted in some cases.