FOLIO-4237 Upgrade to Java 21
This commit bumps the module from Java 17 to 21. It also updates to
dependencies that are compatible with this upgrade and makes a few small
test and config changes to account for these dependeny upgrades. Finally, it
upgrades a few integration tests to use PostgreSQL 16 instead of 12
(this is unreleated to the Java upgrade, but is still important)