MODFQMMGR-412 Fix regresssion
This commit fixes a regression that happens when all fields in a query
are removed during a migration, resulting in the query becoming an empty
string (which isn't valid JSON). We used to set the version property in
those cases, but the removal of version handling in migration strategy
chated this. The fix here is to simply use "{}" instead of an empty
string.