MODFQMMGR-991 Update the available-joins API
This commit changes the overall workflow for dealing with the join
condition portion of the response from this API. Prior to this, users
would select a target ET ID and the API would provide a list fields from
their sources that can be used in a join with the target and a list of
fields from the target that can be used in a join with the sources. With
this commit, the user selects a target ET ID and the API response with a
single list of target and source field pairs, representing a
relationship between the entity types. This simplifies the
implementation and API interactions quite a bit.