MODFQMMGR-982 Add QUEUED query status
This new status is for queries which are queued for execution but
haven't actually started yet, so they get special handling when looking
for zombie queries. A queued query can be queued for 1 hour by default
before it is considered a zombie.
By default, this new status should not leak from the API, and will be
converted to IN_PROGRESS. This behavior can be overridden with the new
CONVERT_QUEUED_STATUS_TO_IN_PROGRESS env var feature flag, so that we
can easily toggle this behavior after updating FQM's consumers.